resturcture folder and use path alias
