+++ components/Maps/GeoJSONLayer/GeoJSON.tsx
@@ -1,49 +1,13 @@
+import { FC } from "react";
 import { Feature } from 'geojson';
-import { geoJSON, FeatureGroup as FeatureGroupLeaflet, StyleFunction } from "leaflet";
-import { createElementObject, createPathComponent, LeafletContextInterface, PathProps, extendContext } from '@react-leaflet/core'
+import { Popup } from "react-leaflet";
+import { GeoJSONPopup } from "./GeoJSONPopup";
+import { GeoJSONComponent } from "./GeoJSONComponent";
 
-interface GeoJSONProps extends PathProps, React.PropsWithChildren {
-  data: Feature;
-}
-
-const createStyles: StyleFunction = () => {
-  const colors = [
-    '#fffddd',
-    '#faf3c8',
-    '#f6e8b3',
-    '#f4dd9f',
-    '#f3d18b',
-    '#f2c578',
-    '#f2b866',
-    '#f2ab55',
-    '#f39d46',
-    '#f38e38',
-    '#f47d2c',
-    '#f56b23',
-    '#f6571d',
-    '#f63c1a',
-    '#f6081b'
-  ]
-
-  return {
-    fillColor: colors[Math.floor(14 * Math.random())],
-    weight: 1,
-    opacity: 1,
-    color: '#888',
-    dashArray: '3',
-    fillOpacity: 0.7
-  }
-}
-
-const createGeoJSON = (props: GeoJSONProps, context: LeafletContextInterface) => {
-  const geoJSONObject = geoJSON(props.data, { style: createStyles() })
-  return createElementObject(geoJSONObject, extendContext(context, { overlayContainer: geoJSONObject }))
-}
-
-const updateGeoJSON = (instance: FeatureGroupLeaflet, props: GeoJSONProps, prevProps: GeoJSONProps) => {
-  if (props.data?.properties !== prevProps.data?.properties) {
-    instance.setStyle(createStyles())
-  }
-}
-
-export const GeoJSON = createPathComponent(createGeoJSON, updateGeoJSON)
+export const GeoJSON: FC<{ feature: Feature }> = ({ feature }) => (
+  <GeoJSONComponent data={feature}>
+    <Popup minWidth={100} closeButton={false}>
+      <GeoJSONPopup properties={feature.properties} />
+    </Popup>
+  </GeoJSONComponent>
+)
