+++ lib/provider/DataProvider.tsx
@@ -1,4 +1,4 @@
-import { GeoJSONProvider } from '../Context/GeoJSONContext';
+import { GeoJSONProvider } from '@context/GeoJSONContext';
 
 export const DataProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => (
   <GeoJSONProvider>
