+++ lib/provider/MapProvider.tsx
@@ -1,5 +1,5 @@
-import { BaseMapProvider } from '../Context/BaseMapContext';
-import { MapSettingsProvider } from '../Context/MapSettingsContext';
+import { BaseMapProvider } from '@context/BaseMapContext';
+import { MapSettingsProvider } from '@context/MapSettingsContext';
 
 export const MapProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => (
   <MapSettingsProvider>
