+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import "../styles/custom-leaflet-popup.css"
+import "@styles/custom-leaflet-popup.css"
 import type { AppProps } from 'next/app'
 
 function MyApp({ Component, pageProps }: AppProps) {
