+++ pages/index.tsx
@@ -1,13 +1,11 @@
+import Head from 'next/head';
 import dynamic from 'next/dynamic';
 import type { NextPage } from 'next'
-import { Data } from '../components/Sidebar/Data';
-import { Options } from '../components/Sidebar/Options';
-import AppProvider from '../components/Provider/AppProvider';
-import { createStyles, ScrollArea, Tabs } from '@mantine/core';
-import { Settings as SettingsPage } from '../components/Sidebar/Settings';
-import { Adjustments, Camera, Settings, Table } from 'tabler-icons-react';
+import { createStyles } from '@mantine/core';
+import { Sidebar } from '@components/Sidebar';
+import AppProvider from '@provider/AppProvider';
 
-const MapWithNoSSR = dynamic(() => import("../components/Maps"), { ssr: false });
+const MapWithNoSSR = dynamic(() => import("@components/Maps"), { ssr: false });
 
 const useStyles = createStyles((theme) => {
   const BREAKPOINT = theme.breakpoints.lg
@@ -35,17 +33,6 @@ const useStyles = createStyles((theme) => {
         maxWidth: theme.breakpoints.xs,
       },
     },
-    tab_body: {
-      padding: theme.spacing.md,
-      height: `calc(100vh - ${40 + 2 * theme.spacing.xs}px)`,
-      display: 'flex',
-      flexDirection: 'column'
-    },
-    scroll_viewport: {
-      '> div': {
-        display: 'block !important'
-      }
-    }
   };
 });
 
@@ -53,37 +40,21 @@ const Home: NextPage = () => {
   const { classes } = useStyles();
 
   return (
-    <AppProvider>
-      <div className={classes.container}>
-        <div className={classes.map_container}>
-          <MapWithNoSSR />
+    <>
+      <Head>
+        <title>App Title</title>
+      </Head>
+      <AppProvider>
+        <div className={classes.container}>
+          <div className={classes.map_container}>
+            <MapWithNoSSR />
+          </div>
+          <div className={classes.sidebar_container}>
+            <Sidebar />
+          </div>
         </div>
-        <div className={classes.sidebar_container}>
-          <Tabs classNames={{ body: classes.tab_body }}>
-            <Tabs.Tab label="Data" icon={<Table size={14} />}>
-              <ScrollArea sx={{ flex: 1 }} type="scroll" scrollHideDelay={200} classNames={{viewport: classes.scroll_viewport}} scrollbarSize={5} >
-                <Data />
-              </ScrollArea>
-            </Tabs.Tab>
-            <Tabs.Tab label="Options" icon={<Adjustments size={14} />}>
-              <ScrollArea sx={{ flex: 1 }} type="scroll" scrollHideDelay={200} scrollbarSize={5} >
-                <Options />
-              </ScrollArea>
-            </Tabs.Tab>
-            <Tabs.Tab label="Screenshoot" icon={<Camera size={14} />}>
-              <ScrollArea sx={{ flex: 1 }} type="scroll" scrollHideDelay={200} scrollbarSize={5} >
-                Screenshoot
-              </ScrollArea>
-            </Tabs.Tab>
-            <Tabs.Tab label="Settings" icon={<Settings size={14} />}>
-              <ScrollArea sx={{ flex: 1 }} type="scroll" scrollHideDelay={200} scrollbarSize={5} >
-                <SettingsPage />
-              </ScrollArea>
-            </Tabs.Tab>
-          </Tabs>
-        </div>
-      </div>
-    </AppProvider>
+      </AppProvider>
+    </>
   )
 }
